+++ DesignToolbox/DesignToolbox/Pages/Components/Chips/ChipElements.swift
@@ -1,0 +1,50 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDS
+import OUDSComponents
+import SwiftUI
+
+struct ChipElements: DesignToolboxElement {
+    let name: String
+    let illustration: AnyView
+    let pageDescription: AnyView
+
+    init() {
+        let variants: [DesignToolboxElement] = [
+            SuggestionChipElement(),
+            FilterChipElement(),
+            ChipPickerElement(),
+        ]
+
+        name = "app_components_chip_label".localized()
+        illustration = AnyView(ChipIllustration())
+        pageDescription = AnyView(DesignToolboxElementPage(
+            name: name,
+            illustration: illustration,
+            description: "app_components_chip_description_text",
+            demoScreen: AnyView(DesignToolboxVariantElement(elements: variants))))
+    }
+}
+
+private struct ChipIllustration: View {
+
+    @Environment(\.theme) private var theme
+
+    var body: some View {
+        VStack(spacing: theme.spaces.spaceFixedSm) {
+            OUDSFilterChip(text: "Filter", selected: true) {}
+            OUDSSuggestionChip(text: "Suggestion") {}
+        }
+    }
+}