+++ DesignToolbox/DesignToolbox/Pages/Components/Chips/Common/ChipLayout.swift
@@ -1,0 +1,33 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+// MARK: - Chip Layout
+
+enum ChipLayout: CaseIterable, CustomStringConvertible {
+    case textOnly
+    case textAndIcon
+    case iconOnly
+
+    var description: String {
+        switch self {
+        case .textOnly:
+            "app_components_common_textOnlyLayout_label"
+        case .textAndIcon:
+            "app_components_common_textAndIconLayout_label"
+        case .iconOnly:
+            "app_components_common_iconOnlyLayout_label"
+        }
+    }
+
+    var id: String { description }
+}