+++ DesignToolbox/DesignToolbox/Pages/Components/Chips/Filter/FilterChipElement.swift
@@ -1,0 +1,32 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDS
+import SwiftUI
+
+struct FilterChipElement: DesignToolboxElement {
+    let name: String
+    let illustration: AnyView
+    let pageDescription: AnyView
+
+    init() {
+        name = "app_components_chip_filterChip_label".localized()
+        illustration = AnyView(EmptyView())
+        pageDescription = AnyView(DesignToolboxElementPage(
+            name: name,
+            description: "app_components_chip_filterChip_description_text",
+            version: OUDSVersions.componentChipVersion,
+            demoScreen: AnyView(FilterChipPage()))
+        )
+    }
+}