+++ DesignToolbox/DesignToolbox/Pages/Components/Radio/Radio/RadioConfiguration.swift
@@ -72,7 +72,7 @@ struct RadioConfiguration: View {
 
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
-            OUDSSwitchItem("app_components_controlItem_selection_label", isOn: $configurationModel.selection)
+            OUDSSwitchItem("app_components_common_selection_label", isOn: $configurationModel.selection)
                 .disabled(!configurationModel.enabled || configurationModel.isError)
 
             OUDSSwitchItem("app_common_enabled_label", isOn: $configurationModel.enabled)